+++ GoRogue.UnitTests/FOV/FOVTests.cs
@@ -9,7 +9,7 @@ using Xunit;
 using Xunit.Abstractions;
 using XUnit.ValueTuples;
 
-namespace GoRogue.UnitTests
+namespace GoRogue.UnitTests.FOV
 {
     public class FOVTests
     {
@@ -18,43 +18,51 @@ namespace GoRogue.UnitTests
         private readonly ITestOutputHelper _output;
 
         // Properties of the test map
-        private const int _height = 30;
-        private const int _width = 30;
-        private static readonly Point _center = (_width / 2, _height / 2);
-        private const int _radius = 10;
+        private const int Height = 30;
+        private const int Width = 30;
+        private static readonly Point s_center = (Width / 2, Height / 2);
+        private const int Radius = 10;
 
         // Basic rectangle LOS map
-        private static readonly IGridView<bool> _losMap = MockMaps.Rectangle(_width, _height);
+        private static readonly IGridView<bool> s_losMap = MockMaps.Rectangle(Width, Height);
 
         // LOS map with double-thick walls
-        private static readonly IGridView<bool> _losMapDoubleThickWalls =
-            MockMaps.DoubleThickRectangle(_width, _height);
+        private static readonly IGridView<bool> s_losMapDoubleThickWalls =
+            MockMaps.DoubleThickRectangle(Width, Height);
 
         // Radius shapes to test
         public static readonly Radius[] Radii = TestUtils.GetEnumValues<Radius.Types>().Select(i => (Radius)i).ToArray();
 
-        [Theory]
-        [MemberDataEnumerable(nameof(Radii))]
-        public void OpenMapEqualToRadius(Radius shape)
+        // Types of FOV to test; must have constructor with 2 params; transparencyView and pointHasher
+        public static readonly Type[] Types =
         {
-            var los = new RecursiveShadowcastingFOV(_losMap);
-            los.Calculate(_center.X, _center.Y, _radius, shape);
+            typeof(RecursiveShadowcastingFOV), typeof(RecursiveShadowcastingDoubleBasedFOV)
+        };
 
-            var radArea = shape.PositionsInRadius(_center, _radius).ToHashSet();
+        public static readonly (Radius radiusShape, Type FOVType)[] TestData = Radii.Combinate(Types).ToArray();
+
+        [Theory]
+        [MemberDataTuple(nameof(TestData))]
+        public void OpenMapEqualToRadius(Radius shape, Type fovType)
+        {
+            var los = (IFOV)Activator.CreateInstance(fovType, s_losMap, null)!;
+            los.Calculate(s_center.X, s_center.Y, Radius, shape);
+
+            var radArea = shape.PositionsInRadius(s_center, Radius).ToHashSet();
             var losArea = los.DoubleResultView.Positions().Where(pos => los.DoubleResultView[pos] > 0.0).ToHashSet();
 
             Assert.Equal(radArea, losArea);
         }
 
         [Theory]
-        [MemberDataEnumerable(nameof(Radii))]
-        public void VisibilityStoppedByWalls(Radius shape)
+        [MemberDataTuple(nameof(TestData))]
+        public void VisibilityStoppedByWalls(Radius shape, Type fovType)
         {
             // FOV over open map
-            var fov = new RecursiveShadowcastingFOV(_losMapDoubleThickWalls);
+            var fov = (IFOV)Activator.CreateInstance(fovType, s_losMapDoubleThickWalls, null)!;
 
             // Calculate LOS with infinite radius
-            fov.Calculate(_center, double.MaxValue, shape);
+            fov.Calculate(s_center, double.MaxValue, shape);
 
             // Verify that the only non-lit positions are the outer walls (which are blocked
             // by the inner ones)
@@ -64,11 +72,11 @@ namespace GoRogue.UnitTests
         }
 
         [Theory]
-        [MemberDataEnumerable(nameof(Radii))]
-        public void BooleanOutput(Radius shape)
+        [MemberDataTuple(nameof(TestData))]
+        public void BooleanOutput(Radius shape, Type fovType)
         {
-            var fov = new RecursiveShadowcastingFOV(_losMap);
-            fov.Calculate(_center, _radius, shape);
+            var fov = (IFOV)Activator.CreateInstance(fovType, s_losMap, null)!;
+            fov.Calculate(s_center, Radius, shape);
 
             _output.WriteLine("FOV for reference:");
             //_output.WriteLine(fov.ToString(2));
@@ -82,30 +90,30 @@ namespace GoRogue.UnitTests
         }
 
         [Theory]
-        [MemberDataEnumerable(nameof(Radii))]
-        public void CurrentHash(Radius shape)
+        [MemberDataTuple(nameof(TestData))]
+        public void CurrentHash(Radius shape, Type fovType)
         {
-            var fov = new RecursiveShadowcastingFOV(_losMap);
+            var fov = (IFOV)Activator.CreateInstance(fovType, s_losMap, null)!;
 
-            fov.Calculate(_center, _radius, shape);
+            fov.Calculate(s_center, Radius, shape);
 
             // Inefficient copy but fine for testing
             var currentFov = new HashSet<Point>(fov.CurrentFOV);
 
-            foreach (var pos in _losMap.Positions())
+            foreach (var pos in s_losMap.Positions())
                 Assert.Equal(fov.DoubleResultView[pos] > 0.0, currentFov.Contains(pos));
         }
 
         [Theory]
-        [MemberDataEnumerable(nameof(Radii))]
-        public void NewlySeenUnseen(Radius shape)
+        [MemberDataTuple(nameof(TestData))]
+        public void NewlySeenUnseen(Radius shape, Type fovType)
         {
-            var fov = new RecursiveShadowcastingFOV(_losMap);
+            var fov = (IFOV)Activator.CreateInstance(fovType, s_losMap, null)!;
 
-            fov.Calculate(_center, _radius, shape);
+            fov.Calculate(s_center, Radius, shape);
             var prevFov = new HashSet<Point>(fov.CurrentFOV);
 
-            fov.Calculate(_center - 1, _radius, shape);
+            fov.Calculate(s_center - 1, Radius, shape);
             var curFov = new HashSet<Point>(fov.CurrentFOV);
             var newlySeen = new HashSet<Point>(fov.NewlySeen);
             var newlyUnseen = new HashSet<Point>(fov.NewlyUnseen);
@@ -117,26 +125,27 @@ namespace GoRogue.UnitTests
                 Assert.NotEqual(prevFov.Contains(pos), newlySeen.Contains(pos));
         }
 
-        [Fact]
-        public void AccessibleBeforeCalculate()
+        [Theory]
+        [MemberDataEnumerable(nameof(Types))]
+        public void AccessibleBeforeCalculate(Type fovType)
         {
-            var fov = new RecursiveShadowcastingFOV(_losMap);
+            var fov = (IFOV)Activator.CreateInstance(fovType, s_losMap, null)!;
             foreach (var pos in fov.DoubleResultView.Positions())
                 Assert.Equal(0.0, fov.DoubleResultView[pos]);
         }
 
         [Theory]
-        [MemberDataEnumerable(nameof(Radii))]
-        public void MultipleFOVDistanceOverlap(Radius shape)
+        [MemberDataTuple(nameof(TestData))]
+        public void MultipleFOVDistanceOverlap(Radius shape, Type fovType)
         {
-            var fov = new RecursiveShadowcastingFOV(_losMap);
-            var decay = 1.0 / (_radius + 1);
+            var fov = (IFOV)Activator.CreateInstance(fovType, s_losMap, null)!;
+            var decay = 1.0 / (Radius + 1);
             Distance dist = shape;
 
             // Calculate two overlapping FOVs
-            var point2 = _center + 5;
-            fov.Calculate(_center, _radius, shape);
-            fov.CalculateAppend(point2, _radius, shape);
+            var point2 = s_center + 5;
+            fov.Calculate(s_center, Radius, shape);
+            fov.CalculateAppend(point2, Radius, shape);
 
             // Print result, for reference
             _output.WriteLine("Resulting FOV (For Reference):");
@@ -146,9 +155,9 @@ namespace GoRogue.UnitTests
             // radii were handled properly)
             foreach (var pos in fov.DoubleResultView.Positions())
             {
-                double minDist = Math.Min(dist.Calculate(pos, _center), dist.Calculate(pos, point2));
+                double minDist = Math.Min(dist.Calculate(pos, s_center), dist.Calculate(pos, point2));
 
-                double expectedVal = minDist > _radius ? 0 : 1.0 - (decay * minDist);
+                double expectedVal = minDist > Radius ? 0 : 1.0 - (decay * minDist);
                 Assert.Equal(expectedVal, fov.DoubleResultView[pos]);
             }
         }