+++ GoRogue/Components/ComponentTagPair.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using System.Runtime.Serialization;
 using JetBrains.Annotations;
@@ -13,7 +12,6 @@ namespace GoRogue.Components
     [DataContract]
     [PublicAPI]
     // Tuples do not resolve names properly; function is provided
-    [SuppressMessage("ReSharper", "CA2225")]
     public readonly struct ComponentTagPair : IEquatable<ComponentTagPair>, IMatchable<ComponentTagPair>
     {
         /// <summary>