+++ GoRogue/SerializedTypes/MapGeneration/RectangleEdgePositionsList.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.Serialization;
 using SadRogue.Primitives.SerializedTypes;
@@ -14,7 +13,6 @@ namespace GoRogue.SerializedTypes.MapGeneration
     /// </summary>
     [PublicAPI]
     [DataContract]
-    [SuppressMessage("ReSharper", "CA1815")] // Type should only be used for serialization
     public struct RectangleEdgePositionsListSerialized
     {
         /// <summary>